requirements.txt
ccxt>=3.0.0,<4.0.0
pandas<2.0.0
numpy<1.25.0
ta==0.11.0
python-dotenv==1.0.0
schedule==1.2.1
matplotlib<3.8.0